Made numerouse changes to begin implementing support for multibodies and user-defined trajectories. Also made some code-style improvements to molecule class. Eliminated support for calculation of radius of gyration from molecule objects in preparation for porting this functionality to multibodies. Molecule class is no longer a child of trajectory, so molecule center of mass analysis has been eliminated. THis functionality will be incorporated into multibodies.
git-svn-id: file:///home/pkawak/git/svn/tmp/AMDAT_trunk/AMDAT/trunk@37 25419694-f2fb-4e8e-af61-79680089b9de

+++ trajectory.cpp
@@ -82,7 +82,7 @@ void Trajectory::clear_memory()
 /*--------------------------------------------------------------------------------------*/
 
 /*Method to define trajectory size and allocate memory appropriately*/
-void Trajectory::set(int atom_type,int timesteps, int m)
+void Trajectory::set(int atom_type,int timesteps, float m)
 {
 	clear_memory();					//clear memory assigned to existing arrays so they may be reassigned
